fix: move PartialJSONSchema to jsonSchemaTypes
